[cli] rename sources to reduce confusion
