+++ etc/visual-studio/libopenthread-cli-windows.vcxproj
@@ -60,7 +60,7 @@
     <ClCompile Include="..\..\src\cli\cli.cpp" />
     <ClCompile Include="..\..\src\cli\cli_dataset.cpp" />
     <ClCompile Include="..\..\src\cli\cli_instance.cpp" />
-    <ClCompile Include="..\..\src\cli\cli_uart.cpp" />
+    <ClCompile Include="..\..\src\cli\uart.cpp" />
     <ClCompile Include="..\..\src\core\utils\missing_strnlen.c" />
     <ClCompile Include="..\..\src\core\utils\parse_cmdline.cpp" />
   </ItemGroup>
