+++ etc/visual-studio/libopenthread-cli-windows.vcxproj.filters
@@ -16,7 +16,7 @@
     <ClCompile Include="..\..\src\cli\cli_instance.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\cli\cli_uart.cpp">
+    <ClCompile Include="..\..\src\cli\uart.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
