+++ etc/visual-studio/libopenthread-cli.vcxproj
@@ -58,7 +58,7 @@
   <ItemGroup>
     <ClCompile Include="..\..\src\cli\cli.cpp" />
     <ClCompile Include="..\..\src\cli\cli_dataset.cpp" />
-    <ClCompile Include="..\..\src\cli\cli_uart.cpp" />
+    <ClCompile Include="..\..\src\cli\uart.cpp" />
     <ClCompile Include="..\..\src\cli\cli_udp_example.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
