+++ etc/visual-studio/libopenthread-cli.vcxproj.filters
@@ -13,7 +13,7 @@
     <ClCompile Include="..\..\src\cli\cli_dataset.cpp">
       <Filter>Source Files</Filter>
     </ClCompile> 
-    <ClCompile Include="..\..\src\cli\cli_uart.cpp">
+    <ClCompile Include="..\..\src\cli\uart.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\cli\cli_udp_example.cpp">
