+++ src/cli/Makefile.am
@@ -145,10 +145,10 @@ SOURCES_COMMON =                      \
     cli.cpp                           \
     cli_coap.cpp                      \
     cli_coap_secure.cpp               \
-    cli_console.cpp                   \
     cli_dataset.cpp                   \
-    cli_uart.cpp                      \
     cli_udp_example.cpp               \
+    uart.cpp                          \
+    console.cpp                       \
     $(NULL)
 
 libopenthread_cli_ftd_a_SOURCES =     \
@@ -163,11 +163,11 @@ noinst_HEADERS                      = \
     cli.hpp                           \
     cli_coap.hpp                      \
     cli_coap_secure.hpp               \
-    cli_console.hpp                   \
     cli_dataset.hpp                   \
     cli_server.hpp                    \
-    cli_uart.hpp                      \
     cli_udp_example.hpp               \
+    console.hpp                       \
+    uart.hpp                          \
     x509_cert_key.hpp                 \
     $(NULL)
