+++ src/cli/uart.hpp
@@ -28,27 +28,27 @@
 
 /**
  * @file
- *   This file contains definitions for a CLI server on the CONSOLE service.
+ *   This file contains definitions for a CLI server on the UART service.
  */
 
-#ifndef CLI_CONSOLE_HPP_
-#define CLI_CONSOLE_HPP_
+#ifndef OPENTHREAD_CLI_UART_HPP_
+#define OPENTHREAD_CLI_UART_HPP_
 
 #include "openthread-core-config.h"
 
-#include <openthread/cli.h>
-
 #include "cli/cli.hpp"
 #include "cli/cli_server.hpp"
+#include "common/instance.hpp"
+#include "common/tasklet.hpp"
 
 namespace ot {
 namespace Cli {
 
 /**
- * This class implements the CLI server on top of the CONSOLE platform abstraction.
+ * This class implements the CLI server on top of the UART platform abstraction.
  *
  */
-class Console : public Server
+class Uart : public Server
 {
 public:
     /**
@@ -57,7 +57,7 @@ public:
      * @param[in]  aInstance  The OpenThread instance structure.
      *
      */
-    Console(Instance *aInstance);
+    Uart(Instance *aInstance);
 
     /**
      * This method delivers raw characters to the client.
@@ -82,36 +82,55 @@ public:
     virtual int OutputFormat(const char *fmt, ...);
 
     /**
-     * This method sets a callback that is called when console has some output.
+     * This method delivers formatted output to the client.
      *
-     * @param[in]  aCallback   A pointer to a callback method.
+     * @param[in]  aFmt  A pointer to the format string.
+     * @param[in]  aAp   A variable list of arguments for format.
+     *
+     * @returns The number of bytes placed in the output queue.
      *
      */
-    void SetOutputCallback(otCliConsoleOutputCallback aCallback);
+    int OutputFormatV(const char *aFmt, va_list aAp);
 
     /**
-     * This method sets a context that is returned with the callback.
+     * This method returns a reference to the interpreter object.
      *
-     * @param[in]  aContext   A pointer to a user context.
+     * @returns A reference to the interpreter object.
      *
      */
-    void SetContext(void *aContext);
+    Interpreter &GetInterpreter(void) { return mInterpreter; }
 
-    void ReceiveTask(char *aBuf, uint16_t aBufLength);
+    void ReceiveTask(const uint8_t *aBuf, uint16_t aBufLength);
+    void SendDoneTask(void);
+
+    static Uart *sUartServer;
 
 private:
     enum
     {
-        kMaxLineLength = 128,
+        kRxBufferSize  = OPENTHREAD_CONFIG_CLI_UART_RX_BUFFER_SIZE,
+        kTxBufferSize  = OPENTHREAD_CONFIG_CLI_UART_TX_BUFFER_SIZE,
+        kMaxLineLength = OPENTHREAD_CONFIG_CLI_MAX_LINE_LENGTH,
     };
 
-    otCliConsoleOutputCallback mCallback;
-    void *                     mContext;
+    otError ProcessCommand(void);
+    void    Send(void);
+
+    char     mRxBuffer[kRxBufferSize];
+    uint16_t mRxLength;
+
+    char     mTxBuffer[kTxBufferSize];
+    uint16_t mTxHead;
+    uint16_t mTxLength;
+
+    uint16_t mSendLength;
 
     Interpreter mInterpreter;
+
+    friend class Interpreter;
 };
 
 } // namespace Cli
 } // namespace ot
 
-#endif // CLI_CONSOLE_HPP_
+#endif // OPENTHREAD_CLI_UART_HPP_
